pyproject.toml
[tool.poetry]
name = "simple-health-checker"
version = "0.1.0"
description = ""
authors = ["Kallil Souza dos Santos <kallilsouza@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
python-dotenv = "^1.1.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
